+++ example/spec/parallel/createReconciliationReport/CreateReconciliationReportSpec.js
@@ -169,7 +169,6 @@ describe('When there are granule differences and granule reconciliation is run',
 
     config = await loadConfig();
 
-    process.env.CollectionsTable = `${config.stackName}-CollectionsTable`;
     process.env.GranulesTable = `${config.stackName}-GranulesTable`;
     granuleModel = new Granule();