+++ tf-modules/cumulus/ingest.tf
@@ -40,9 +40,7 @@ module "ingest" {
   custom_queues = var.custom_queues
   throttled_queues = var.throttled_queues
 
-  report_executions_sns_topic_arn = module.archive.report_executions_sns_topic_arn
-  report_granules_sns_topic_arn   = module.archive.report_granules_sns_topic_arn
-  report_pdrs_sns_topic_arn       = module.archive.report_pdrs_sns_topic_arn
+  sf_event_sqs_to_db_records_sqs_queue_url = module.archive.sf_event_sqs_to_db_records_sqs_queue_url
 
   tags = var.tags
 }